+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/dto/StripeDtoRequest.kt
@@ -3,7 +3,7 @@ package com.example.eccomerce_app.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class StripeDtoRequest(val amount: Double, val currency:String)
+data class StripeDtoRequest(val amount: Long, val currency:String)
 
 @Serializable
 data class StripeClientSecret(val client_secret:String)
