+++ eccorce_app/app/src/main/java/com/example/eccomerce_app/model/CartModel.kt
@@ -1,9 +1,10 @@
 package com.example.e_commercompose.model
 
+import com.example.eccomerce_app.model.CardProductModel
 import java.util.UUID
 
 data class CartModel(
-    val totalPrice: Double,
+    val totalPrice: Long,
     val longitude: Double,
     val latitude: Double,
     val userId: UUID,
